Determine required samples in strata

Determine required samples in strata

calculate_allocation(
  sraster,
  nSamp,
  allocation = "prop",
  mraster = NULL,
  existing = NULL,
  force = FALSE
)

Arguments

sraster

spatRaster. Stratification raster to be used for sampling.

nSamp

Numeric. Number of desired samples.

allocation

Character. Allocation algorithm to be used. Either prop (default) for proportional allocation or optim for optimal allocation (equal sampling cost) or equal for equal number of samples (defined by nSamp) for each strata.

mraster

spatRaster. ALS metric raster. Required when allocation = optim.

existing

sf or data.frame. Existing plot network.

force

Logical. Default = FALSE - force nSamp to be exactly the user defined value in cases where nSamp and sraster strata count are not equally divisible. Has no effect when existing is provided.

Value

data.frame of:

strata - Strata ID.

total - Total samples to be allocated (positive) or removed (negative).

need - Total required samples per strata.

Details

Determine how many samples to allocate within strata

References

Gregoire, T.G., & Valentine, H.T. (2007). Sampling Strategies for Natural Resources and the Environment (1st ed.). Chapman and Hall/CRC. https://doi.org/10.1201/9780203498880

See also

Author

Tristan R.H. Goodbody

Examples

#--- Load strata raster and existing samples---#
r <- system.file("extdata", "kmeans.tif", package = "sgsR")
sr <- terra::rast(r)

e <- system.file("extdata", "existing.shp", package = "sgsR")
e <- sf::st_read(e)
#> Reading layer `existing' from data source 
#>   `C:\Users\tgood.stu\AppData\Local\Temp\RtmpwjzfyE\temp_libpath29801e1871d0\sgsR\extdata\existing.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431110 ymin: 5337750 xmax: 438550 ymax: 5343230
#> CRS:           unknown

#--- perform grid sampling ---#
calculate_allocation(
  sraster = sr,
  nSamp = 200
)
#> Implementing porportional allocation of samples
#> nSamp of 200 is not perfectly divisible based on strata distribution. nSamp of 199 will be returned. Use "force = TRUE" to brute force to 200.
#>    strata total
#> 1       1    13
#> 2       2    29
#> 3       3    14
#> 4       4    21
#> 5       5    25
#> 6       6    24
#> 7       7    24
#> 8       8    16
#> 9       9    23
#> 10     10    10

calculate_allocation(
  sraster = sr,
  nSamp = 200,
  force = TRUE
)
#> Implementing porportional allocation of samples
#> Forcing 200 total samples.
#>    strata total
#> 1       1    13
#> 2       2    29
#> 3       3    14
#> 4       4    21
#> 5       5    25
#> 6       6    24
#> 7       7    24
#> 8       8    16
#> 9       9    23
#> 10     10    11

#--- extract strata from existing samples ---#
e.sr <- extract_strata(
  sraster = sr,
  existing = e
)

calculate_allocation(
  sraster = sr,
  nSamp = 200,
  existing = e.sr
)
#> Implementing porportional allocation of samples
#> nSamp of 200 is not perfectly divisible based on strata distribution. nSamp of 199 will be returned. Use "force = TRUE" to brute force to 200.
#>    strata total need
#> 1       1    11   13
#> 2       2    15   29
#> 3       3     2   14
#> 4       4    10   21
#> 5       5    10   25
#> 6       6    10   24
#> 7       7    12   24
#> 8       8     3   16
#> 9       9    18   23
#> 10     10     8   10

#--- Load mraster for optimal allocation ---#
mr <- system.file("extdata", "wall_metrics.tif", package = "sgsR")
mr <- terra::rast(mr)

calculate_allocation(
  sraster = sr,
  nSamp = 200,
  existing = e.sr,
  allocation = "optim",
  mraster = mr$zmax,
  force = TRUE
)
#> Implementing optimal allocation of samples based on variability of 'zmax'
#>    strata total need
#> 1       1    16   18
#> 2       2    15   29
#> 3       3     6   18
#> 4       4     7   18
#> 5       5     6   21
#> 6       6     9   23
#> 7       7    12   24
#> 8       8     6   19
#> 9       9     9   14
#> 10     10    14   16