Sample size determination

Determine a samples size for simple random sampling using relative standard error

calculate_sampsize(
  mraster,
  rse = NULL,
  start = 0.01,
  end = 0.05,
  increment = 0.001,
  plot = FALSE
)

Arguments

mraster

spatRaster. Metrics raster. All values must be numeric.

rse

Numeric. Desired relative standard error (coefficient of variation of the mean) threshold to determine sample size.

start

Numeric. First rse value to begin rse sequence. default = 0.01.

end

Numeric. Final rse value to end rse sequence. default = 0.05.

increment

Numeric. Value to increment between start & end. default = 0.001.

plot

Logical. if TRUE output graphical representation of estimated sample size vs. rse.

Value

A data.frame of sample size and rse by raster variable.

Note

$$rse = (100 * SE) / mean)$$

Where:

SE - Standard error of the mean

s - Standard deviation of the observations

n - Number of observations

References

Benedetti, R., Piersimoni, F., & Postiglione, P. (2015). Sampling spatial units for agricultural surveys. pp 202-203. Berlin: Springer.

See also

Author

Tristan R.H. Goodbody

Examples

#--- Load raster ---#
r <- system.file("extdata", "wall_metrics.tif", package = "sgsR")
mr <- terra::rast(r)

calculate_sampsize(mraster = mr,
                   rse = 0.01,
                   plot = TRUE)
                   
calculate_sampsize(mraster = mr,
                   plot = TRUE)
                   
calculate_sampsize(mraster = mr,
                   rse = 0.025,
                   start = 0.01,
                   end = 0.08,
                   increment = 0.01,
                   plot = TRUE)
                   
#--- higher variance leads to need for more samples ---#