Sampling rules

Sampling rules

Stratified rule 1

Stratified rule 2

AHELS nSamp

AHELS threshold

Usage

strat_rule1(n, s, i, strat_mask, add_strata, extraCols, mindist)

strat_rule2(n, s, add_strata, nCount, strata_m, extraCols, mindist)

ahels_nSamp(
  nSamp,
  nQuant,
  tolerance,
  nb,
  underRep,
  ratio,
  ratOrderUnder,
  matCovDens,
  matCovSampDens,
  samples,
  mats,
  vals
)

ahels_threshold(
  threshold,
  tolerance,
  nQuant,
  nb,
  ratio,
  underRep,
  ratOrderUnder,
  matCovDens,
  matCovSampDens,
  samples,
  mats,
  vals
)

Arguments

mindist

Numeric. Minimum allowable distance between selected samples. Default = NULL.

nSamp

Numeric. Number of desired samples. existing, include and force influence this value.

nQuant

Numeric. Number of quantiles to divide covariates and samples into. Quantiles that do not cover at least 1 percent of the area of interest will be excluded and be returned as NA.

tolerance

Numeric. Allowable tolerance (<= 0.1 (10 added until the 1 - tolerance density is reached. If threshold is used, samples will be added until the threshold - tolerance value is reached. This parameter allows the user to define a buffer around desired quantile densities to permit the algorithm to not add additional samples if quantile density is very close to 1, or user-defined threshold.

threshold

Numeric. Sample quantile ratio threshold. After the threshold default = 0.9 is reached, no additional samples will be added. Values close to 1 can cause the algorithm to continually loop.

Value

Allocated samples to be returned by sample_strat() & sample_ahels().