Balanced sampling

Balanced raster sampling using lpm2 and lpm2_kdtree methods

sample_balanced(
  mraster,
  nSamp,
  algorithm = "lpm2",
  p = NULL,
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  plot = FALSE,
  filename = NULL,
  overwrite = FALSE
)

Arguments

mraster

spatRaster. ALS metrics raster.

nSamp

Numeric. Number of desired samples.

algorithm

Character. One of lpm2 lcube lcubestratified

p

Numeric. Vector with length equal to the number of cells in mraster representing the inclusion probability for each candidate sample. Default = nSamp / N, where N is the number of cells.

access

sf. Road access network - must be lines.

buff_inner

Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.

buff_outer

Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.

plot

Logical. Plots output strata raster and visualized strata with boundary dividers.

filename

Character. Path to write stratified raster to disc.

overwrite

Logical. Specify whether filename should be overwritten on disc.

Value

An sf object with nSamp randomly sampled points.

References

Anton Grafstrom and Jonathan Lisic (2019). BalancedSampling: Balanced and Spatially Balanced Sampling. R package version 1.5.5. https://CRAN.R-project.org/package=BalancedSampling

Jonathan Lisic and Anton Grafstrom (2018). SamplingBigData: Sampling Methods for Big Data. R package version 1.0.0. https://CRAN.R-project.org/package=SamplingBigData

See also

Author

Tristan R.H. Goodbody

Examples

if (FALSE) {
#--- Load raster and existing plots---#
r <- system.file("extdata", "wall_metrics.tif", package = "sgsR")
mr <- terra::rast(r)

a <- system.file("extdata", "roads.shp", package = "sgsR")
ac <- sf::st_read(a)

sample_balanced(
  mraster = mr,
  nSamp = 200,
  plot = TRUE
)

sample_balanced(
  mraster = mr,
  nSamp = 100,
  algorithm = "lcube",
  access = ac,
  buff_inner = 50,
  buff_outer = 200
)
}