Sample existing

Sub-sample an existing sample. Four sampling methods are available: clhs, balanced, srs and strat.

Usage

sample_existing(
  existing,
  nSamp,
  raster = NULL,
  type = "clhs",
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)

Arguments

existing

sf 'POINT'. Existing plot network.

nSamp

Numeric. Number of desired samples.

raster

SpatRaster. Raster to guide the location of the samples. If type = "clhs" this raster can also be used to define the population distributions to be used for sampling.

type

Character. A string indicating the type of sampling method to use. Possible values are "clhs", "balanced", "srs" and "strat".

access

sf. Road access network - must be lines.

buff_inner

Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.

buff_outer

Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.

details

Logical. If FALSE (default) output is sf object of systematic samples. If TRUE returns a list of sf objects where tessellation is the tessellation grid for sampling, and samples are the systematic samples.

filename

Character. Path to write output samples.

overwrite

Logical. Choice to overwrite existing filename if it exists.

...

Additional arguments for the sampling method selected.

Value

An sf object of samples or a list object if details = TRUE

Note

When type = "clhs" or type = "balanced" all attributes in existing will be used for sampling. Remove attributes not indented for sampling' prior to using this algorithm.

See also

Author

Tristan R.H. Goodbody

Examples

r <- system.file("extdata", "mraster.tif", package = "sgsR")
mr <- terra::rast(r)

#--- generate an existing sample adn extract metrics ---#
e <- sample_systematic(raster = mr, cellsize = 200)
e <- extract_metrics(existing = e, mraster = mr)

#--- perform clhs (default) sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50
)
#> Sub-sampling based on ALL 'existing' metric distributions. Ensure only attributes of interest are included.
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431120.3 ymin: 5337764 xmax: 438501.6 ymax: 5343217
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 889  8.02     32.6 2.13 POINT (437680.8 5338372)
#> 757  8.78     83.1 1.85 POINT (437761.4 5341092)
#> 629 14.50     90.4 3.39 POINT (434278.1 5338223)
#> 339  3.24      8.6 0.65 POINT (433955.9 5340748)
#> 346 22.50     94.1 4.99 POINT (432854.5 5339325)
#> 268 19.50     98.8 3.88 POINT (434006.6 5341468)
#> 32  15.70     88.2 3.64 POINT (432021.8 5342498)
#> 502 26.30     95.2 6.55 POINT (432911.1 5337764)
#> 501 23.90     91.3 5.41 POINT (433033.5 5337922)
#> 891  6.73     83.8 1.51   POINT (437436 5338056)

#--- perform balanced sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "balanced"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431114.4 ymin: 5337779 xmax: 438450.8 ymax: 5343193
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 8   19.60     88.0 4.13 POINT (431547.3 5342865)
#> 18  13.10     88.7 3.28 POINT (431215.9 5342110)
#> 30  13.40     68.9 3.16 POINT (432266.6 5342814)
#> 70  10.80     60.6 2.92 POINT (431114.4 5340671)
#> 95  16.90     64.6 4.31   POINT (432777 5342166)
#> 124 21.30     90.4 5.23 POINT (431833.7 5340620)
#> 129 19.50     85.9 4.22 POINT (431221.8 5339829)
#> 143 18.00     94.0 3.46 POINT (432481.4 5341131)
#> 146  9.09     80.1 2.33 POINT (432114.3 5340656)
#> 156  9.49     22.6 2.31 POINT (434230.6 5343065)

#--- perform simple random sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "srs"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431135.3 ymin: 5337850 xmax: 438415 ymax: 5343151
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 646 20.20     85.0 6.13 POINT (436639.2 5340948)
#> 295  5.08     67.4 1.12 POINT (434776.7 5342136)
#> 907  6.90      8.3 1.79 POINT (437997.1 5338128)
#> 118 11.30     80.9 2.30   POINT (432568 5341569)
#> 506 15.80     84.8 4.27 POINT (436618.3 5342229)
#> 383 17.00     91.4 4.85 POINT (432767.9 5338886)
#> 637  3.20     16.3 0.96   POINT (437863 5342530)
#> 140 15.80     79.7 4.39 POINT (432848.6 5341605)
#> 119 20.10     88.0 5.24 POINT (432445.6 5341411)
#> 747 10.90     66.0 2.88 POINT (435400.4 5338367)