sgsR: Structurally Guided Sampling

Structurally guided sampling (SGS) approaches for airborne laser scanning (ALS; LIDAR). Primary functions provide means to generate data-driven stratifications & methods for allocating samples. Intermediate functions for calculating and extracting important information about input covariates and samples are also included. Processing outcomes are intended to help forest and environmental management practitioners better optimize field sample placement as well as assess and augment existing sample networks in the context of data distributions and conditions. ALS data is the primary intended use case, however any rasterized remote sensing data can be used, enabling data-driven stratifications and sampling approaches.

See also

Author

Maintainer: Tristan RH Goodbody goodbody.t@gmail.com (ORCID) [copyright holder]

Authors:

Nicholas C Coops nicholas.coops@ubc.ca (ORCID)

Martin Queinnec queinnec@mail.ubc.ca (ORCID)