Sample allocation type and count

Determine how many samples to allocate within strata.

calculate_allocation(
  sraster,
  nSamp,
  allocation = "prop",
  weights = NULL,
  mraster = NULL,
  existing = NULL,
  force = FALSE
)

Arguments

sraster

spatRaster. Stratification raster to be used for sampling.

nSamp

Numeric. Number of desired samples.

allocation

Character. Allocation algorithm to be used. Either prop (default) for proportional allocation, optim for optimal allocation (equal sampling cost), equal for equal number of samples (defined by nSamp) for each strata, or "manual" for user defined strata weights defined using weights.

weights

Numeric. Only applicable when allocation = "manual". Vector of weights where sum(weights) == 1. Vector length must be equal to the number of unique strata where the first numeric value corresponds to stratum 1, second stratum 2 etc.

mraster

spatRaster. ALS metric raster. Required when allocation = optim.

existing

sf 'POINT' or data.frame. Existing plot network.

force

Logical. Default = FALSE - force nSamp to be exactly the user defined value in cases where nSamp and sraster strata count are not equally divisible. Additional samples often need to be allocated or removed based on rounding differences resulting from proportional differences between nSamp and strata coverages in sraster. In these instances samples are either added to strata with the lowest number of samples or are removed from strata with the highest number of samples. Has no effect when existing is provided.

Value

Returns a data.frame of:

strata - Strata ID.

total - Total samples to be allocated based on under representation (positive) or over representation (negative) to be removed at the users discretion.

need - Total required samples per strata. Rounded.

References

Gregoire, T.G., & Valentine, H.T. (2007). Sampling Strategies for Natural Resources and the Environment (1st ed.). Chapman and Hall/CRC. https://doi.org/10.1201/9780203498880

See also

Author

Tristan R.H. Goodbody

Examples

#--- Load strata raster and existing samples---#
r <- system.file("extdata", "sraster.tif", package = "sgsR")
sr <- terra::rast(r)

e <- system.file("extdata", "existing.shp", package = "sgsR")
e <- sf::st_read(e)
#> Reading layer `existing' from data source 
#>   `/home/runner/work/_temp/Library/sgsR/extdata/existing.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 200 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431110 ymin: 5337710 xmax: 438530 ymax: 5343230
#> Projected CRS: UTM_Zone_17_Northern_Hemisphere

#--- proportional allocation ---#
calculate_allocation(
  sraster = sr,
  nSamp = 50
)
#> Implementing proportional allocation of samples.
#> nSamp of 50 is not perfectly divisible based on strata distribution. nSamp of 51 will be returned. Use 'force = TRUE' to brute force to 50.
#>   strata total
#> 1      1    13
#> 2      2    13
#> 3      3    13
#> 4      4    12

#--- equal allocation ---#
calculate_allocation(
  sraster = sr,
  allocation = "equal",
  nSamp = 10
)
#> Implementing equal allocation of samples.
#> # A tibble: 4 × 2
#>   strata total
#>    <dbl> <dbl>
#> 1      1    10
#> 2      2    10
#> 3      3    10
#> 4      4    10

#--- manual allocation ---#
#--- define user-defined weights ---#

weights <- c(0.2, 0.2, 0.5, 0.1)

calculate_allocation(
  sraster = sr,
  allocation = "manual",
  weights = weights,
  nSamp = 200
)
#> Implementing allocation of samples based on user-defined weights.
#>   strata total
#> 1      1    40
#> 2      2    40
#> 3      3   100
#> 4      4    20