Analyze optimal Latin hypercube sample number

Population level analysis of metric raster data to determine optimal Latin Hypercube sample size

calculate_lhsOpt(
  popLHS,
  PCA = TRUE,
  quant = TRUE,
  KLdiv = TRUE,
  minSamp = 10,
  maxSamp = 100,
  step = 10,
  rep = 10,
  iter = 10000
)

Arguments

popLHS

List. Output from calculate_lhsPop function.

PCA

Logical. Calculates principal component loadings of the population for PCA similarity factor testing. default = TRUE.

quant

Logical. Calculates quantile matrix of the population for quantile comparison testing. default = TRUE.

KLdiv

Logical. Calculates covariate matrix of the population for Kullback–Leibler divergence testing. default = TRUE. Relies on quant = TRUE to calculate.

minSamp

Numeric. Minimum sample size to test. default = 10.

maxSamp

Numeric. Maximum sample size to test. default = 100.

step

Numeric. Sample step size for each iteration. default = 10.

rep

Numeric. Internal repetitions for each sample size. default = 10.

iter

Numeric. Internal to clhs - A positive number, giving the number of iterations for the Metropolis-Hastings annealing process. Defaults to 10000.

Value

data.frame with summary statistics.

Note

Special thanks to Dr. Brendan Malone for the original implementation of this algorithm.

References

Malone BP, Minansy B, Brungard C. 2019. Some methods to improve the utility of conditioned Latin hypercube sampling. PeerJ 7:e6451 DOI 10.7717/peerj.6451

Author

Tristan R.H. Goodbody

Examples

if (FALSE) {
#--- Load raster and access files ---#
r <- system.file("extdata", "wall_metrics.tif", package = "sgsR")
mr <- terra::rast(r)

#--- calculate lhsPop details ---#
poplhs <- calculate_lhsPop(mraster = mr)

calculate_lhsOpt(popLHS = poplhs)

calculate_lhsOpt(
  popLHS = poplhs,
  PCA = FALSE,
  iter = 200
)
}