Compare sample representation within sraster strata

Compare sample representation within sraster strata

calculate_representation(sraster, existing, plot = FALSE)

Arguments

sraster

spatRaster. Stratification raster.

existing

sf 'POINT'. Existing plot network.

plot

Logical. Plot frequency of strata coverage and sampling coverage for sraster and existing. Will return a list if TRUE.

Value

Returns a tibble where:

strata - sraster strata ID.

srasterFreq - sraster coverage frequency.

sampleFreq - Sampling frequency within sraster strata.

diffFreq - Difference between srasterFreq & sampleFreq. Positive values indicate over representation.

nSamp - Number of samples within each strata in existing.

need - srasterFreq * sum(nSamp). Total theoretical number of required samples to be representative of strata coverage. This values is rounded. It is important for the user to consider diffFreq. A small difference - e.g. 1 in sampleFreq vs. srasterFreq frequency could make the algorithm allocate or remove samples that could likely be ignored.

Details

Calculate how sraster strata are represented in existing samples

See also

Author

Tristan R.H. Goodbody, Martin Queinnec

Examples

### --- generate example stratification ---###

#--- load ALS metrics from sgsR internal data ---#
r <- system.file("extdata", "mraster.tif", package = "sgsR")

#--- read ALS metrics using the terra package ---#
mraster <- terra::rast(r)

#--- perform stratification ---#
sraster <- strat_kmeans(
  mraster = mraster$zq90,
  nStrata = 6
)
#> K-means being performed on 1 layers with 6 centers.

### --- create existing sample network ---###

#--- simple random sampling ---#
existing <- sample_srs(
  raster = mraster$zq90,
  nSamp = 100
)

#--- calculate representation ---#

calculate_representation(
  sraster = sraster,
  existing = existing
)
#> # A tibble: 6 × 6
#>   strata srasterFreq sampleFreq diffFreq nSamp  need
#>    <dbl>       <dbl>      <dbl>    <dbl> <int> <dbl>
#> 1      1        0.18       0.2    0.0200    20    -2
#> 2      2        0.09       0.09   0          9     0
#> 3      3        0.09       0.12   0.03      12    -3
#> 4      4        0.12       0.07  -0.05       7     5
#> 5      5        0.24       0.24   0         24     0
#> 6      6        0.27       0.28   0.0100    28    -1