Sample existing

Sub-sample an existing sample. Four sampling methods are available: clhs, balanced, srs and strat.

Usage

sample_existing(
  existing,
  nSamp,
  raster = NULL,
  type = "clhs",
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)

Arguments

existing

sf 'POINT'. Existing plot network.

nSamp

Numeric. Number of desired samples.

raster

SpatRaster. Raster to guide the location of the samples. If type = "clhs" this raster can also be used to define the population distributions to be used for sampling.

type

Character. A string indicating the type of sampling method to use. Possible values are "clhs", "balanced", "srs" and "strat".

access

sf. Road access network - must be lines.

buff_inner

Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.

buff_outer

Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.

details

Logical. If FALSE (default) output is sf object of systematic samples. If TRUE returns a list of sf objects where tessellation is the tessellation grid for sampling, and samples are the systematic samples.

filename

Character. Path to write output samples.

overwrite

Logical. Choice to overwrite existing filename if it exists.

...

Additional arguments for the sampling method selected.

Value

An sf object of samples or a list object if details = TRUE

Note

When type = "clhs" or type = "balanced" all attributes in existing will be used for sampling. Remove attributes not indented for sampling' prior to using this algorithm.

See also

Author

Tristan R.H. Goodbody

Examples

r <- system.file("extdata", "mraster.tif", package = "sgsR")
mr <- terra::rast(r)

#--- generate an existing sample adn extract metrics ---#
e <- sample_systematic(raster = mr, cellsize = 200)
e <- extract_metrics(existing = e, mraster = mr)

#--- perform clhs (default) sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50
)
#> Sub-sampling based on ALL 'existing' metric distributions. Ensure only attributes of interest are included.
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431434.6 ymin: 5337704 xmax: 438495.9 ymax: 5343069
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 425  6.47     23.2 1.87 POINT (434251.9 5340639)
#> 765 11.40     66.2 3.16 POINT (434648.9 5342689)
#> 494  3.11     14.6 0.54 POINT (433551.9 5341458)
#> 793  6.90     49.0 1.67   POINT (435096 5342681)
#> 326 11.30     73.9 3.07 POINT (435567.8 5339272)
#> 610 18.90     93.6 4.25 POINT (433653.2 5342082)
#> 906  4.52     41.3 1.03 POINT (437517.3 5342993)
#> 612 17.90     92.1 4.10 POINT (434007.5 5341897)
#> 648 15.30     87.8 3.57 POINT (434454.7 5341888)
#> 422  2.26      0.9 0.32 POINT (433720.5 5340918)

#--- perform balanced sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "balanced"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431518.9 ymin: 5337738 xmax: 438546.1 ymax: 5343170
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>     zq90 pzabove2  zsd                 geometry
#> 4   16.4     92.5 2.88 POINT (431602.3 5337738)
#> 14  17.8     97.9 3.50   POINT (431788 5338092)
#> 52  15.5     78.3 4.43 POINT (432159.3 5338801)
#> 93  12.3     74.4 3.39 POINT (434647.9 5337948)
#> 148 15.4     94.9 3.00 POINT (432092.1 5339965)
#> 156 17.1     90.8 3.65 POINT (433686.4 5339129)
#> 194 17.7     96.2 3.63 POINT (435196.4 5338564)
#> 212 20.5     93.4 4.80 POINT (432986.3 5339948)
#> 244 22.4     65.7 7.16 POINT (433079.2 5340125)
#> 248 13.3     97.5 2.60 POINT (433787.8 5339754)

#--- perform simple random sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "srs"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431493.4 ymin: 5337804 xmax: 438520.5 ymax: 5343145
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>     zq90 pzabove2  zsd                 geometry
#> 1  25.80     91.9 7.09 POINT (433306.5 5337974)
#> 2   7.51     91.9 1.45 POINT (438023.2 5341373)
#> 3  13.60     87.5 3.03 POINT (431493.4 5341408)
#> 4  11.50     79.2 2.33   POINT (434826 5342597)
#> 5  12.90     60.9 3.72 POINT (432184.9 5340142)
#> 6  15.00     62.0 4.46 POINT (437027.6 5340765)
#> 7  16.80     87.4 4.23 POINT (432590.4 5342639)
#> 8  15.60     85.2 4.21 POINT (437053.2 5342107)
#> 9  14.90     90.5 2.77 POINT (432379.1 5340944)
#> 10 17.30     82.1 4.34 POINT (432919.1 5341112)