Sample existing

Sub-sample an existing sample. Four sampling methods are available: clhs, balanced, srs and strat.

Usage

sample_existing(
  existing,
  nSamp,
  raster = NULL,
  type = "clhs",
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)

Arguments

existing

sf 'POINT'. Existing plot network.

nSamp

Numeric. Number of desired samples.

raster

SpatRaster. Raster to guide the location of the samples. If type = "clhs" this raster can also be used to define the population distributions to be used for sampling.

type

Character. A string indicating the type of sampling method to use. Possible values are "clhs", "balanced", "srs" and "strat".

access

sf. Road access network - must be lines.

buff_inner

Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.

buff_outer

Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.

details

Logical. If FALSE (default) output is sf object of systematic samples. If TRUE returns a list of sf objects where tessellation is the tessellation grid for sampling, and samples are the systematic samples.

filename

Character. Path to write output samples.

overwrite

Logical. Choice to overwrite existing filename if it exists.

...

Additional arguments for the sampling method selected.

Value

An sf object of samples or a list object if details = TRUE

Note

When type = "clhs" or type = "balanced" all attributes in existing will be used for sampling. Remove attributes not indented for sampling' prior to using this algorithm.

See also

Author

Tristan R.H. Goodbody

Examples

r <- system.file("extdata", "mraster.tif", package = "sgsR")
mr <- terra::rast(r)

#--- generate an existing sample adn extract metrics ---#
e <- sample_systematic(raster = mr, cellsize = 200)
e <- extract_metrics(existing = e, mraster = mr)

#--- perform clhs (default) sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50
)
#> Sub-sampling based on ALL 'existing' metric distributions. Ensure only attributes of interest are included.
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431705 ymin: 5337855 xmax: 438180.5 ymax: 5343170
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 658  3.42      3.2 0.69 POINT (434341.4 5338918)
#> 488  6.04     68.5 1.22 POINT (432690.4 5343170)
#> 311  5.60     13.0 1.55 POINT (434487.6 5342757)
#> 496 15.40      7.8 4.20 POINT (435383.3 5339191)
#> 395 15.70     83.3 3.19 POINT (433886.3 5342562)
#> 501 20.30     41.2 6.57 POINT (434669.9 5340156)
#> 485 14.40     61.5 3.34 POINT (433047.1 5342687)
#> 702 21.70     87.6 4.55 POINT (432634.8 5340890)
#> 267 11.10     84.7 2.77 POINT (437026.5 5339660)
#> 595 16.80     97.0 3.10 POINT (434306.3 5339639)

#--- perform balanced sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "balanced"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431278.1 ymin: 5337744 xmax: 438348.1 ymax: 5343094
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 10   4.22      6.8 0.94   POINT (437970 5343094)
#> 17   2.75     12.7 0.45 POINT (437809.2 5342975)
#> 78  23.50     95.2 6.05 POINT (436529.5 5343024)
#> 100 17.40     87.4 5.07 POINT (438348.1 5339891)
#> 102 12.80     33.1 3.50 POINT (438110.3 5340213)
#> 105 18.30     91.3 4.59 POINT (437753.6 5340695)
#> 213 17.70     85.3 4.46 POINT (436515.8 5341024)
#> 220  5.34     65.5 1.29 POINT (435564.6 5342310)
#> 228  3.19      2.7 0.57 POINT (438257.5 5338332)
#> 235 10.60     62.0 2.91 POINT (437425.2 5339457)

#--- perform simple random sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "srs"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431208 ymin: 5337821 xmax: 438376.4 ymax: 5343198
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>     zq90 pzabove2  zsd                 geometry
#> 1  10.40     52.5 2.72 POINT (438264.3 5339332)
#> 2   7.40     35.4 1.88 POINT (436096.7 5338226)
#> 3  19.60     85.1 6.18 POINT (434040.3 5341680)
#> 4   9.33     37.8 2.50   POINT (437740 5338695)
#> 5  15.70     82.2 3.83 POINT (433229.4 5340086)
#> 6   7.62     35.6 2.11 POINT (431522.8 5342058)
#> 7  19.50     98.4 3.98 POINT (431586.1 5339617)
#> 8  19.60     87.0 5.46   POINT (431208 5342820)
#> 9  13.50     97.7 2.81   POINT (436894 5337821)
#> 10  3.42     19.4 0.65 POINT (434194.2 5340799)