Sample existing

Sub-sample an existing sample. Four sampling methods are available: clhs, balanced, srs and strat.

Usage

sample_existing(
  existing,
  nSamp,
  raster = NULL,
  type = "clhs",
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)

Arguments

existing

sf 'POINT'. Existing plot network.

nSamp

Numeric. Number of desired samples.

raster

SpatRaster. Raster to guide the location of the samples. If type = "clhs" this raster can also be used to define the population distributions to be used for sampling.

type

Character. A string indicating the type of sampling method to use. Possible values are "clhs", "balanced", "srs" and "strat".

access

sf. Road access network - must be lines.

buff_inner

Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.

buff_outer

Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.

details

Logical. If FALSE (default) output is sf object of systematic samples. If TRUE returns a list of sf objects where tessellation is the tessellation grid for sampling, and samples are the systematic samples.

filename

Character. Path to write output samples.

overwrite

Logical. Choice to overwrite existing filename if it exists.

...

Additional arguments for the sampling method selected.

Value

An sf object of samples or a list object if details = TRUE

Note

When type = "clhs" or type = "balanced" all attributes in existing will be used for sampling. Remove attributes not indented for sampling' prior to using this algorithm.

See also

Author

Tristan R.H. Goodbody

Examples

r <- system.file("extdata", "mraster.tif", package = "sgsR")
mr <- terra::rast(r)

#--- generate an existing sample adn extract metrics ---#
e <- sample_systematic(raster = mr, cellsize = 200)
e <- extract_metrics(existing = e, mraster = mr)

#--- perform clhs (default) sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50
)
#> Sub-sampling based on ALL 'existing' metric distributions. Ensure only attributes of interest are included.
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431584.7 ymin: 5337726 xmax: 438221.6 ymax: 5343124
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 683 21.60     90.8 7.18 POINT (437776.4 5340602)
#> 386  3.68     15.8 0.82 POINT (433754.2 5340455)
#> 37  16.90     65.3 4.85   POINT (431630 5342257)
#> 371 11.60     92.6 2.71 POINT (436913.4 5342181)
#> 765  9.37     24.4 2.53 POINT (433755.9 5337949)
#> 735 29.00     95.3 7.25 POINT (433835.5 5338220)
#> 422 21.30     94.9 5.10 POINT (432972.5 5339800)
#> 303 14.80     84.3 3.63   POINT (435493 5341861)
#> 261  1.62      0.0 0.12 POINT (436450.2 5342612)
#> 307  4.92      5.6 2.61 POINT (434790.9 5341477)

#--- perform balanced sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "balanced"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431103.4 ymin: 5337708 xmax: 438559.8 ymax: 5343231
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 1   18.90     81.5 4.65   POINT (431326 5343231)
#> 40  15.90     73.6 4.53 POINT (431103.4 5341970)
#> 70  17.80     93.6 3.66 POINT (431470.7 5341714)
#> 86  15.60     90.3 3.57 POINT (431917.7 5341731)
#> 100 16.10     29.8 3.95 POINT (433242.1 5342226)
#> 105 15.50     87.3 3.09 POINT (432189.1 5341651)
#> 121 16.70     86.4 3.73 POINT (433689.1 5342243)
#> 131 20.90     84.7 5.33 POINT (431933.9 5341284)
#> 148 17.40     90.6 2.94 POINT (433784.9 5342067)
#> 170  5.04     52.8 1.09 POINT (435109.4 5342563)

#--- perform simple random sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "srs"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431121.5 ymin: 5337886 xmax: 438429.6 ymax: 5343231
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>    zq90 pzabove2  zsd                 geometry
#> 1  11.7     83.0 3.33 POINT (438047.8 5340522)
#> 2  20.5     91.6 5.19 POINT (431425.4 5338043)
#> 3  13.9     92.4 3.41 POINT (434950.2 5342020)
#> 4  14.3     95.2 2.52 POINT (433768.7 5342514)
#> 5  18.0     94.3 4.57 POINT (431662.5 5341363)
#> 6  16.4     97.9 2.51 POINT (431121.5 5339017)
#> 7  21.7     92.6 5.05 POINT (433260.2 5339274)
#> 8   8.6     39.3 2.28 POINT (432125.7 5340933)
#> 9  17.4     90.4 4.13 POINT (433162.5 5341955)
#> 10 15.2     55.1 4.70 POINT (433195.1 5341061)