Sample existing

Sub-sample an existing sample. Four sampling methods are available: clhs, balanced, srs and strat.

Usage

sample_existing(
  existing,
  nSamp,
  raster = NULL,
  type = "clhs",
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)

Arguments

existing

sf 'POINT'. Existing plot network.

nSamp

Numeric. Number of desired samples.

raster

SpatRaster. Raster to guide the location of the samples. If type = "clhs" this raster can also be used to define the population distributions to be used for sampling.

type

Character. A string indicating the type of sampling method to use. Possible values are "clhs", "balanced", "srs" and "strat".

access

sf. Road access network - must be lines.

buff_inner

Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.

buff_outer

Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.

details

Logical. If FALSE (default) output is sf object of systematic samples. If TRUE returns a list of sf objects where tessellation is the tessellation grid for sampling, and samples are the systematic samples.

filename

Character. Path to write output samples.

overwrite

Logical. Choice to overwrite existing filename if it exists.

...

Additional arguments for the sampling method selected.

Value

An sf object of samples or a list object if details = TRUE

Note

When type = "clhs" or type = "balanced" all attributes in existing will be used for sampling. Remove attributes not indented for sampling' prior to using this algorithm.

See also

Author

Tristan R.H. Goodbody

Examples

r <- system.file("extdata", "mraster.tif", package = "sgsR")
mr <- terra::rast(r)

#--- generate an existing sample adn extract metrics ---#
e <- sample_systematic(raster = mr, cellsize = 200)
e <- extract_metrics(existing = e, mraster = mr)

#--- perform clhs (default) sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50
)
#> Sub-sampling based on ALL 'existing' metric distributions. Ensure only attributes of interest are included.
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431160.2 ymin: 5337701 xmax: 438528.6 ymax: 5343008
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 414  8.75     24.4 2.59 POINT (438448.4 5339221)
#> 192 24.30     97.9 5.51 POINT (433114.2 5339381)
#> 74  16.70     89.0 3.86 POINT (435568.3 5337703)
#> 290 14.70     85.9 3.38 POINT (434040.3 5339758)
#> 106 16.20     90.9 4.15   POINT (433929 5338547)
#> 773 18.90     86.8 4.31   POINT (437409 5341562)
#> 238 22.60     92.6 7.95 POINT (435292.8 5339016)
#> 10  11.30     72.7 3.32 POINT (432270.1 5337742)
#> 214 10.30     48.0 2.90 POINT (438143.6 5338060)
#> 30  19.90     95.9 3.06 POINT (432758.5 5338027)

#--- perform balanced sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "balanced"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431209.1 ymin: 5337753 xmax: 438468 ymax: 5343037
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 14  19.70     91.0 6.24 POINT (431740.5 5338088)
#> 18  13.80     98.7 2.99 POINT (432514.3 5337884)
#> 28  13.30     91.5 2.90 POINT (432371.7 5338129)
#> 73  11.10     89.4 1.76 POINT (435374.9 5337753)
#> 97  15.70     88.6 4.28 POINT (431220.8 5339258)
#> 108 17.50     76.2 5.19 POINT (434315.8 5338445)
#> 114 15.70     86.9 3.21 POINT (435476.5 5338140)
#> 131 14.50     80.5 3.84 POINT (433979.8 5338740)
#> 220 17.80     92.2 4.68 POINT (431810.9 5339930)
#> 271  7.05     38.2 1.79 POINT (437471.5 5338650)

#--- perform simple random sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "srs"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431138.7 ymin: 5337734 xmax: 438458.2 ymax: 5343168
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 818  9.76     69.5 2.52 POINT (434415.5 5342761)
#> 400  2.20      0.1 0.27 POINT (435546.8 5339983)
#> 141 10.10     91.7 2.41   POINT (436301 5338131)
#> 753 18.60     80.0 4.70 POINT (433540.2 5342578)
#> 822 12.20     67.5 3.43 POINT (435189.3 5342558)
#> 338  2.32      2.1 0.31 POINT (437766.5 5338986)
#> 78  22.10     85.4 5.68 POINT (432330.6 5338760)
#> 923 10.80     66.6 2.65 POINT (438202.3 5343008)
#> 466  3.09      7.3 0.57 POINT (433907.5 5340827)
#> 759  5.89     45.2 1.42 POINT (434700.8 5342273)