Sample existing

Sub-sample an existing sample. Four sampling methods are available: clhs, balanced, srs and strat.

Usage

sample_existing(
  existing,
  nSamp,
  raster = NULL,
  type = "clhs",
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)

Arguments

existing

sf 'POINT'. Existing plot network.

nSamp

Numeric. Number of desired samples.

raster

SpatRaster. Raster to guide the location of the samples. If type = "clhs" this raster can also be used to define the population distributions to be used for sampling.

type

Character. A string indicating the type of sampling method to use. Possible values are "clhs", "balanced", "srs" and "strat".

access

sf. Road access network - must be lines.

buff_inner

Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.

buff_outer

Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.

details

Logical. If FALSE (default) output is sf object of systematic samples. If TRUE returns a list of sf objects where tessellation is the tessellation grid for sampling, and samples are the systematic samples.

filename

Character. Path to write output samples.

overwrite

Logical. Choice to overwrite existing filename if it exists.

...

Additional arguments for the sampling method selected.

Value

An sf object of samples or a list object if details = TRUE

Note

When type = "clhs" or type = "balanced" all attributes in existing will be used for sampling. Remove attributes not indented for sampling' prior to using this algorithm.

See also

Author

Tristan R.H. Goodbody

Examples

r <- system.file("extdata", "mraster.tif", package = "sgsR")
mr <- terra::rast(r)

#--- generate an existing sample adn extract metrics ---#
e <- sample_systematic(raster = mr, cellsize = 200)
e <- extract_metrics(existing = e, mraster = mr)

#--- perform clhs (default) sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50
)
#> Sub-sampling based on ALL 'existing' metric distributions. Ensure only attributes of interest are included.
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431135.3 ymin: 5337701 xmax: 438336.3 ymax: 5343123
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 886 14.30     85.0 4.02 POINT (438163.4 5341827)
#> 475 17.90     94.7 3.18 POINT (435694.4 5337992)
#> 1   18.20     95.7 3.38 POINT (431135.3 5338299)
#> 644 11.70     69.8 3.15 POINT (435613.8 5343051)
#> 566  8.01     75.1 1.98   POINT (436383 5338446)
#> 235 18.70     22.7 5.25 POINT (432196.9 5343123)
#> 134 13.40     71.9 3.47 POINT (431972.2 5341348)
#> 44  15.30     84.7 3.83 POINT (431416.2 5340517)
#> 884 21.20     93.2 3.72 POINT (438030.9 5342204)
#> 887 20.60     68.4 6.36   POINT (438296 5341449)

#--- perform balanced sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "balanced"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431227.5 ymin: 5337707 xmax: 438530.8 ymax: 5342888
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 4   15.30     93.3 4.40 POINT (431334.1 5337733)
#> 30  10.10     76.9 2.65 POINT (431227.5 5340451)
#> 31  20.00     89.2 5.01 POINT (431293.7 5340262)
#> 66  14.20     93.9 3.53 POINT (432333.8 5338508)
#> 122 15.10     90.2 3.12   POINT (432910 5338074)
#> 147 16.20     86.0 4.80 POINT (432966.1 5338518)
#> 150 24.90     92.4 6.58 POINT (433164.9 5337952)
#> 156 19.10     86.3 4.31 POINT (431829.6 5342358)
#> 166 18.10     91.4 3.38 POINT (432492.2 5340471)
#> 209  7.12     17.6 1.89 POINT (433674.9 5337707)

#--- perform simple random sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "srs"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431141 ymin: 5337767 xmax: 438530.8 ymax: 5343143
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>    zq90 pzabove2  zsd                 geometry
#> 1  15.7     94.2 3.67 POINT (433343.5 5338650)
#> 2  14.0     87.0 3.75 POINT (436107.8 5339833)
#> 3  29.5     88.9 8.00 POINT (436842.5 5341363)
#> 4  12.1     90.8 3.23 POINT (434680.3 5342087)
#> 5  18.5     57.4 4.41 POINT (433461.7 5343143)
#> 6  18.3     79.4 4.74 POINT (436393.1 5337814)
#> 7  10.3     67.1 2.68 POINT (434104.1 5342521)
#> 8  14.2     59.4 4.42 POINT (432247.3 5339961)
#> 9  14.8     49.6 4.65   POINT (434058 5341445)
#> 10 13.0     78.7 3.13 POINT (433349.3 5342256)