Sample existing

Sub-sample an existing sample using clhs functionality.

Usage

sample_existing(
  existing,
  nSamp,
  raster = NULL,
  cost = NULL,
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  iter = 10000,
  plot = FALSE,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)

Arguments

existing

sf 'POINT'. Existing plot network.

nSamp

Numeric. Number of desired samples.

raster

spatRaster. Raster used to define population distributions.

cost

Numeric/Character. Index or name of covariate within mraster to be used to constrain cLHS sampling. If default (NULL), a cost constraint is not used.

access

sf. Road access network - must be lines.

buff_inner

Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.

buff_outer

Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.

iter

Numeric. Value giving the number of iterations within the Metropolis-Hastings process.

plot

Logical. Plots output strata raster with samples.

details

Logical. If FALSE (default) output is sf object of systematic samples. If TRUE returns a list of sf objects where tessellation is the tessellation grid for sampling, and samples are the systematic samples.

filename

Character. Path to write output samples.

overwrite

Logical. Choice to overwrite existing filename if it exists.

...

Additional arguments for clhs sampling. See clhs.

Value

An sf object of samples or a list object if `details = TRUE`

Note

If providing only existing - all attributes will be used for sampling. Remove attributes not indented for sampling prior to using this algorithm.

See also

Author

Tristan R.H. Goodbody

Examples

#--- Load raster ---#
r <- system.file("extdata", "mraster.tif", package = "sgsR")
mr <- terra::rast(r)

#--- generate an existing sample ---#
e <- sample_systematic(raster = mr, cellsize = 200)

#--- perform sub-sampling ---#
sample_existing(
  existing = e,
  raster = mr,
  nSamp = 50
)
#> 'existing' does not contain attributes with the same names as 'raster'. Extracting metrics.
#> Sub-sampling based on 'raster' distributions.
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431345.6 ymin: 5337725 xmax: 438496.5 ymax: 5343185
#> CRS:           +proj=utm +zone=17 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 551  2.99     10.2 0.54 POINT (434476.1 5341384)
#> 415 15.00     56.1 4.32 POINT (435016.6 5340223)
#> 581  4.06     30.7 0.86 POINT (434296.9 5341473)
#> 33  10.60     96.6 2.18 POINT (437439.3 5337906)
#> 378 16.70     95.0 2.90 POINT (434218.5 5338163)
#> 51   2.64     15.2 0.40 POINT (437793.2 5339070)
#> 89  18.40     94.3 3.16 POINT (438145.6 5340681)
#> 733 11.70     86.5 3.12 POINT (433845.2 5342813)
#> 380 17.00     87.3 3.83 POINT (434396.2 5338522)
#> 609  7.52     26.2 2.07 POINT (434028.8 5341382)