Sample existing

Sub-sample an existing sample. Four sampling methods are available: clhs, balanced, srs and strat.

Usage

sample_existing(
  existing,
  nSamp,
  raster = NULL,
  type = "clhs",
  access = NULL,
  buff_inner = NULL,
  buff_outer = NULL,
  details = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)

Arguments

existing

sf 'POINT'. Existing plot network.

nSamp

Numeric. Number of desired samples.

raster

SpatRaster. Raster to guide the location of the samples. If type = "clhs" this raster can also be used to define the population distributions to be used for sampling.

type

Character. A string indicating the type of sampling method to use. Possible values are "clhs", "balanced", "srs" and "strat".

access

sf. Road access network - must be lines.

buff_inner

Numeric. Inner buffer boundary specifying distance from access where plots cannot be sampled.

buff_outer

Numeric. Outer buffer boundary specifying distance from access where plots can be sampled.

details

Logical. If FALSE (default) output is sf object of systematic samples. If TRUE returns a list of sf objects where tessellation is the tessellation grid for sampling, and samples are the systematic samples.

filename

Character. Path to write output samples.

overwrite

Logical. Choice to overwrite existing filename if it exists.

...

Additional arguments for the sampling method selected.

Value

An sf object of samples or a list object if details = TRUE

Note

When type = "clhs" or type = "balanced" all attributes in existing will be used for sampling. Remove attributes not indented for sampling' prior to using this algorithm.

See also

Author

Tristan R.H. Goodbody

Examples

r <- system.file("extdata", "mraster.tif", package = "sgsR")
mr <- terra::rast(r)

#--- generate an existing sample adn extract metrics ---#
e <- sample_systematic(raster = mr, cellsize = 200)
e <- extract_metrics(existing = e, mraster = mr)

#--- perform clhs (default) sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50
)
#> Sub-sampling based on ALL 'existing' metric distributions. Ensure only attributes of interest are included.
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431305.6 ymin: 5337746 xmax: 438535.2 ymax: 5343171
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 666 20.30     85.9 4.49 POINT (436221.4 5339946)
#> 551 16.30     83.1 5.01 POINT (436108.4 5340569)
#> 622  8.77     94.8 1.61 POINT (437879.1 5341083)
#> 785 10.10     84.2 2.51 POINT (436956.7 5339437)
#> 900  2.35      1.3 0.34 POINT (438076.5 5338226)
#> 153 13.80     75.4 3.42 POINT (433201.1 5341712)
#> 193 15.40     80.6 3.82 POINT (436374.8 5342995)
#> 522 13.90     92.5 3.22 POINT (433907.7 5339590)
#> 48  13.40     67.3 3.15 POINT (432369.7 5342397)
#> 287 14.50     88.5 2.60 POINT (432357.9 5340337)

#--- perform balanced sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "balanced"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431452.3 ymin: 5338007 xmax: 438535.2 ymax: 5343098
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>      zq90 pzabove2  zsd                 geometry
#> 7   20.80     75.4 5.26 POINT (431985.1 5343098)
#> 15  17.40     70.7 4.45 POINT (432256.7 5343019)
#> 17  15.20     50.5 3.40 POINT (431905.9 5342827)
#> 21  10.80     63.8 2.77 POINT (432703.5 5343036)
#> 58   4.79     33.0 1.06 POINT (433693.5 5342894)
#> 66  16.70     72.7 3.28 POINT (432290.4 5342125)
#> 80  16.20     89.1 3.19 POINT (433263.4 5342430)
#> 90  17.40     92.5 3.91 POINT (431509.6 5341469)
#> 128  9.61      3.9 2.93 POINT (433104.9 5341887)
#> 153 13.80     75.4 3.42 POINT (433201.1 5341712)

#--- perform simple random sub-sampling ---#
sample_existing(
  existing = e,
  nSamp = 50,
  type = "srs"
)
#> Simple feature collection with 50 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 431288.7 ymin: 5337735 xmax: 438037.6 ymax: 5343115
#> Projected CRS: UTM Zone 17, Northern Hemisphere
#> First 10 features:
#>     zq90 pzabove2  zsd                 geometry
#> 1  20.60     94.9 5.29   POINT (431481 5339857)
#> 2   5.74     41.5 1.30 POINT (436312.5 5342277)
#> 3  15.80     77.8 4.47 POINT (436742.5 5342741)
#> 4  13.70     68.3 3.43 POINT (435695.3 5339658)
#> 5   7.08     56.8 1.55 POINT (437182.7 5338192)
#> 6   9.50     85.3 2.35   POINT (434621 5342034)
#> 7  11.40     57.2 3.31 POINT (433064.5 5338216)
#> 8  19.00     80.4 5.33 POINT (433958.3 5338250)
#> 9  14.30     91.7 3.38 POINT (433924.6 5339144)
#> 10 19.00     85.9 4.77 POINT (433415.3 5338408)