Map 2 stratified rasters

Map stratified rasters to a combined stratification.

Usage

strat_map(
  sraster,
  sraster2,
  stack = FALSE,
  filename = NULL,
  overwrite = FALSE,
  plot = FALSE,
  details = FALSE
)

Arguments

sraster

spatRaster. Primary stratification raster.

sraster2

spatRaster. Secondary stratification raster.

stack

Logical. Default = FALSE. If TRUE, output raster will be 3 layers: strata, strata2, stratamapped.

filename

Character. Path to write stratified raster to disc.

overwrite

Logical. Specify whether filename should be overwritten on disc.

plot

Logical. Plots output strata raster and visualized strata with boundary dividers.

details

Logical. If FALSE (default) output is a mapped stratified spatRaster object. If TRUE return a list where $outRaster is the mapped stratified raster, and $lookUp is the lookup table for the stratification.

Value

A spatRaster object.

Mapping

The mapping algorithm will take the stratification from sraster and combine it with overlying strata values in sraster2. This will result in a stratamapped attribute where the values from both inputs are combined.

i.e.

If strata = 1 and strata2 = 1 then stratamapped = 11.

If strata = 2 and strata2 = 14 then stratamapped = 214.

See also

Author

Tristan R.H. Goodbody, Robert Hijmans

Examples

#--- load input metrics raster ---#
raster <- system.file("extdata", "sraster.tif", package = "sgsR")
sraster <- terra::rast(raster)

#--- read polygon coverage ---#
poly <- system.file("extdata", "inventory_polygons.shp", package = "sgsR")
fri <- sf::st_read(poly)
#> Reading layer `inventory_polygons' from data source 
#>   `/home/runner/work/_temp/Library/sgsR/extdata/inventory_polygons.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 632 features and 3 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 431100 ymin: 5337700 xmax: 438560 ymax: 5343240
#> Projected CRS: UTM_Zone_17_Northern_Hemisphere

#--- stratify polygon coverage ---#
#--- specify polygon attribute to stratify ---#

attribute <- "NUTRIENTS"

#--- specify features within attribute & how they should be grouped ---#
#--- as a single vector ---#

features <- c("poor", "rich", "medium")

srasterfri <- strat_poly(
  poly = fri,
  attribute = attribute,
  features = features,
  raster = sraster
)

#--- map srasters ---#
strat_map(
  sraster = srasterfri,
  sraster2 = sraster
)
#> class       : SpatRaster 
#> dimensions  : 277, 373, 1  (nrow, ncol, nlyr)
#> resolution  : 20, 20  (x, y)
#> extent      : 431100, 438560, 5337700, 5343240  (xmin, xmax, ymin, ymax)
#> coord. ref. : UTM Zone 17, Northern Hemisphere 
#> source      : memory 
#> categories  : label 
#> name        : strata 
#> min value   :     11 
#> max value   :     34 

strat_map(
  sraster = srasterfri,
  sraster2 = sraster,
  stack = TRUE,
  details = TRUE
)
#> Stacking sraster, sraster2, and their combination (stratamapped).
#> $raster
#> class       : SpatRaster 
#> dimensions  : 277, 373, 3  (nrow, ncol, nlyr)
#> resolution  : 20, 20  (x, y)
#> extent      : 431100, 438560, 5337700, 5343240  (xmin, xmax, ymin, ymax)
#> coord. ref. : UTM Zone 17, Northern Hemisphere 
#> sources     : memory  
#>               sraster.tif  
#>               memory  
#> names       : strata, strata2, stratamapped 
#> min values  :      1,       1,           11 
#> max values  :      3,       4,           34 
#> 
#> $lookUp
#>    strata strata2 stratamapped
#> 1       3       2           32
#> 2       3       1           31
#> 3       1       3           13
#> 4       1       4           14
#> 5       3       4           34
#> 6       3       3           33
#> 7       1       2           12
#> 8       1       1           11
#> 9       2       2           22
#> 10      2       3           23
#> 11      2       4           24
#> 12      2       1           21
#>